requirements.txt
h5py==3.7.0
keras==2.11.0
mediapipe==0.9.0.1
numpy==1.21.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.3.0
streamlit==1.16.0
tensorflow==2.11.0
